+++ YAM2E/FORMS/ImportExportData.Designer.cs
@@ -1,0 +1,48 @@
+﻿namespace LAMP.FORMS
+{
+    partial class ImportExportData
+    {
+        /// <summary>
+        /// Required designer variable.
+        /// </summary>
+        private System.ComponentModel.IContainer components = null;
+
+        /// <summary>
+        /// Clean up any resources being used.
+        /// </summary>
+        /// <param name="disposing">true if managed resources should be disposed; otherwise, false.</param>
+        protected override void Dispose(bool disposing)
+        {
+            if (disposing && (components != null))
+            {
+                components.Dispose();
+            }
+            base.Dispose(disposing);
+        }
+
+        #region Windows Form Designer generated code
+
+        /// <summary>
+        /// Required method for Designer support - do not modify
+        /// the contents of this method with the code editor.
+        /// </summary>
+        private void InitializeComponent()
+        {
+            System.ComponentModel.ComponentResourceManager resources = new System.ComponentModel.ComponentResourceManager(typeof(ImportExportData));
+            SuspendLayout();
+            // 
+            // ImportExportData
+            // 
+            AutoScaleDimensions = new System.Drawing.SizeF(7F, 15F);
+            AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
+            ClientSize = new System.Drawing.Size(824, 476);
+            Icon = (System.Drawing.Icon)resources.GetObject("$this.Icon");
+            Name = "ImportExportData";
+            StartPosition = System.Windows.Forms.FormStartPosition.CenterScreen;
+            Text = "Data Im-/Exporter";
+            ResumeLayout(false);
+        }
+
+        #endregion
+    }
+}